% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColValid.R
\name{ColValid}
\alias{ColValid}
\title{Validation of the Number of Columns}
\usage{
ColValid(x, n = NULL, verbose = TRUE)
}
\arguments{
\item{x}{A matrix-like R object}

\item{n}{An integer representing expected columns}

\item{verbose}{A logical print statements if TRUE}
}
\value{
The number of columns in object if no expected no is sent

TRUE if number of columns matches expected

FALSE if number of columns does not match expected
}
\description{
Validates the number of columns of a loaded data object and
compares it with the expected number of columns provided by the user. If the
user does not provide the expected column parameter, then the function returns the
number of columns.

If Verbose = FALSE, returns dataframe of column count and logical
}
\examples{
df<- data.frame(col1 = c(1,2,3),
               col2 = c(1,2,3),
               col3 = c(1,2,3))

ColValid(df,n = 3)
ColValid(df, n = 2)
cf <- ColValid(df, n = 3, verbose = FALSE)
}
\seealso{
\code{\link{ncol}}
}
