% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountVal.R
\name{CountVal}
\alias{CountVal}
\title{Count of Unique & Duplicated Values of Each Column}
\usage{
CountVal(x, verbose = TRUE)
}
\arguments{
\item{x}{A matrix-like R object}

\item{verbose}{Logical value to print results to console}
}
\value{
If Verbose = TRUE, nothing returned

If Verbose = FALSE, returns data frame of unique & dup values by column
}
\description{
Counts the number of unique and duplicated values by
column. Missing values (NA's) are omitted in the counts and TRUE
values are counted for Factors.

If Verbose = T, displays results to console

If Verbose = F, returns dataframe of counts
}
\examples{
\dontrun{
#EXAMPLE
}
df_type <- data.frame(num1 = rnorm(5),
           int2= c(1:5),
           log3 = c(TRUE, TRUE, FALSE, FALSE, TRUE),
           fac4 = letters[1:5])
CountVal(df_type)
value <- CountVal(df_type, verbose = FALSE)
}
\seealso{
\code{\link{unique, duplicated}}
}
