% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateHux.R
\name{CreateHux}
\alias{CreateHux}
\title{Creates Hux Table Object for Printing}
\usage{
CreateHux(x, title)
}
\arguments{
\item{x}{A matrix-like R object}

\item{title}{String containing plot title}
}
\value{
A hux object
}
\description{
Internal Helper Function for Outlier
}
