% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA.R
\name{EDA}
\alias{EDA}
\title{Performs Multiple EDA Analyis Functions}
\usage{
EDA(x, expR = NULL, expC = NULL, plots = FALSE)
}
\arguments{
\item{x}{A matrix-like R Object}

\item{expR}{Interger of expected row number}

\item{expC}{Integer of expected column number}

\item{plots}{Logical value to produce plots}
}
\value{
A list of ggplot2 plots
}
\description{
Performs multiple exploratory functions on dataframe

Provides:

For Object - name, object class, and random sample of values

For Rows - missing values, percent of total row, and duplicate rows

For Columns - value type, missing values (count & percent),
range (min, max, & spread), Shapiro-Wilk Normality (p-value and 95% determination),
and count any outliers (inner and outer)
}
\examples{
\dontrun{
x <- data.frame("X1" = c(1,2,3,1,2,3),
                "X2" = 1:6,
                "X3"= c("a","b","c","a","b","c"))

EDA(x)

plot_value <- EDA(x,plots = TRUE)
}
}
