% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindColNA.R
\name{FindColNA}
\alias{FindColNA}
\title{Identifies Number and Percentage of NAs By Column}
\usage{
FindColNA(x, verbose = TRUE)
}
\arguments{
\item{x}{A matrix-like R object}

\item{verbose}{Logical value to print results to console}
}
\value{
The NA column details

If verbose is TRUE (default) the results are printed to screen

If verbose is FALSE results are returned via dataframe
}
\description{
Identifies Number and Percentage of NAs By Column
}
\examples{
\dontrun{
#EXAMPLE
}
x = matrix(rnorm(100), ncol = 10)
FindColNA(x)
}
