% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindRowNA.R
\name{FindRowNA}
\alias{FindRowNA}
\title{Identifies Number and Percentage of NAs By Row}
\usage{
FindRowNA(x, verbose = TRUE)
}
\arguments{
\item{x}{A matrix-like R object}

\item{verbose}{Logical value to print results to console}
}
\value{
The NA Row details

If verbose is TRUE (default) the results are printed to screen

If verbose is FALSE results are returned via dataframe
}
\description{
Identifies Number and Percentage of NAs By Row
}
\examples{
\dontrun{
#EXAMPLE
}
x = matrix(rnorm(100), ncol = 10)
x[2,3] <-NA #Adding Missing Value

FindRowNA(x)
}
