% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HasNA.R
\name{HasNA}
\alias{HasNA}
\title{Checks for Missing Values in Dataframe}
\usage{
HasNA(x)
}
\arguments{
\item{x}{a matrix-like R object}
}
\value{
missing_values Logical indicator if NA exist in object

TRUE if  object contains any NA value

FALSE if object does not contain an NA value
}
\description{
Validates if entire object has missing values
}
\examples{
m1 = matrix(rnorm(100), ncol = 10)

HasNA(m1)

}
\seealso{
\code{\link{is.na}}
}
