% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQR.R
\name{IQR}
\alias{IQR}
\title{Calculates Quartiles and Outlier Ranges}
\usage{
IQR(v, fence = 1)
}
\arguments{
\item{v}{Vector}

\item{fence}{A int of either 1 (outer) or 2 (inner)}
}
\value{
Number of outliers
}
\description{
Internal Helper Function for Outlier

Calls OutsideRange
}
