% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQRange.R
\name{IQRange}
\alias{IQRange}
\title{Calculates Quartiles and Outlier Ranges}
\usage{
IQRange(v, fence = 1)
}
\arguments{
\item{v}{Vector}

\item{fence}{A int of either 1 (outer) or 2 (inner)}
}
\value{
Number of outliers
}
\description{
Internal Helper Function for Outlier

Calls OutsideRange
}
