% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiplot.R
\name{Multiplot}
\alias{Multiplot}
\title{Prints Multiple Plots from a List}
\usage{
Multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{Additional grid parameters}

\item{plotlist}{List of ggplot objects}

\item{file}{A file object}

\item{cols}{Int for number of columns to display}

\item{layout}{a grid object}
}
\value{
A list of plots
}
\description{
Function from From: R Cookbook

http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
}
