% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Out.R
\name{Out}
\alias{Out}
\title{Determines Inner and Outer IQR Fence Outlier Counts}
\usage{
Out(x, verbose = TRUE)
}
\arguments{
\item{x}{A matrix-like R object}

\item{verbose}{Logical operator}
}
\value{
If verbose = FALSE, a dataframe of results
}
\description{
Determines number of values by column that fall

inner and outer IQR Range Fences

If verbose = TRUE, displays results to console

If Verbose = FALSE, returns dataframe of results
}
\examples{
\dontrun{
x <- data.frame("X1" = c(71,70,73,70,70,69,70,72,71,300,71,69),
                "X2" = c(71,70,73,70,70,66,70,72,71,300,71,301),
                "X3" = c(71,70,73,70,70,69,70,72,71,71,71,69))

Out(x)

value <- Out(x, verbose = FALSE)
}
}
