% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutsideRange.R
\name{OutsideRange}
\alias{OutsideRange}
\title{Determines if a Value is Outside Range}
\usage{
OutsideRange(x, upper, lower)
}
\arguments{
\item{x}{A vector}

\item{upper}{Numeric top bound}

\item{lower}{Numeric bottom bound}
}
\value{
Numeric
}
\description{
Determines if a Value is Outside Range
}
\details{
Internal Help Function for Outlier and IQR
}
