% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{Plots}
\alias{Plots}
\title{Creates a List of ggplots: density or boxplot}
\usage{
Plots(x, column, colorData = NULL, ptype = 1)
}
\arguments{
\item{x}{A matrix-like R structure}

\item{column}{chr of column name}

\item{colorData}{dataframe column to designate target}

\item{ptype}{int of plot type}
}
\value{
ggplot object
}
\description{
Function to generate ggplot objects using lapply

If the data frame contains a single factor column
the plot will assume it as the dependent variable

If ptype = 1, density plots will be produced

If ptype = 2, box plots will be produced
}
\examples{
\dontrun{
#EXAMPLE
}
x <- data.frame("X1" = c(1,2,3,1,2,3),
                "X2" = 1:6,
                "F3"= c("male","female","male","female","male","male"))

plot_density <- lapply(colnames(x), Plots, x = x, colorData = x$F3, ptype = 1 )
}
