% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rng.R
\name{Rng}
\alias{Rng}
\title{Identifies Max, Min, and Spread of Numeric Columns}
\usage{
Rng(x, verbose = TRUE)
}
\arguments{
\item{x}{A matrix-like R object}

\item{verbose}{Logical operator}
}
\value{
If verbose = FALSE, a dataframe of results
}
\description{
Determines maximum value and minimum value for each column

If verbose = TRUE, displays results to console

If Verbose = FALSE, returns dataframe of results
}
\examples{
\dontrun{
x <- data.frame("X1" = c(1,2,3,1,2,3),
                "X2" = 1:6,
                "X3"= c("a","b","c","a","b","c"))

Rng(x)

value <- Rng(x, verbose = FALSE)
}
}
\seealso{
\code{\link{range}}
}
