% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RowSample.R
\name{RowSample}
\alias{RowSample}
\title{Random Row Sample}
\usage{
RowSample(x, n = 5)
}
\arguments{
\item{x}{a matrix-like R object}

\item{n}{the number of random rows to display}
}
\value{
A dataframe
}
\description{
Validates that the number of columns is equal to expected count
If object row count exceeds supplied parameter or default value,
then the function returns a random sample of "n" rows. If object row
count is less than parameter or default, the function returns all
rows of the object.
}
\examples{
\dontrun{
#Obtain a Random Sample of Rows
}
x = matrix(rnorm(100), ncol = 10)
RowSample(x)
}
\seealso{
\code{\link{sample, head, tail}}
}
