% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColType.R
\name{ColType}
\alias{ColType}
\title{Determining Class Types of Each Column in an Object}
\usage{
ColType(x, verbose = TRUE)
}
\arguments{
\item{x}{A matrix-like R object}

\item{verbose}{Logical value to print results to console}
}
\value{
If verbose = FALSE, a dataframe of column object types
}
\description{
Displays the class of each column in a matrix or dataframe

If Verbose = F, returns dataframe col names and class types
}
\examples{
df<- data.frame(col1 = c(1,2,3),
               col2 = c(1,2,3),
               col3 = c(1,2,3))

ColValid(df,n = 3)
ColValid(df, n = 2)
cf <- ColValid(df, n = 3, verbose = FALSE)
}
\seealso{
\code{\link{class}}
}
