% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColType.R
\name{ColType}
\alias{ColType}
\title{Determining Class Types of Each Column in an Object}
\usage{
ColType(x, verbose = TRUE)
}
\arguments{
\item{x}{A matrix-like R object}

\item{verbose}{Logical value to print results to console}
}
\value{
If verbose = FALSE, a dataframe of column object types
}
\description{
Displays the class of each column in a matrix

or dataframe

If Verbose = F, returns dataframe col names and class types
}
\examples{
/dontrun{
#EXAMPLE
}
x <- data.frame(num1 = rnorm(5),
           int2= c(1:5),
           log3 = c(TRUE, TRUE, FALSE, FALSE, TRUE),
           fac4 = letters[1:5])
ColType(x)
}
\seealso{
\code{\link{typeof,class}}
}
