% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColValid.R
\name{ColValid}
\alias{ColValid}
\title{Validation of the Number of Columns}
\usage{
ColValid(x, n = NULL, verbose = TRUE)
}
\arguments{
\item{x}{a matrix-like R object}

\item{n}{number of expected rows}

\item{verbose}{displays print statements if TRUE}
}
\value{
The number of columns in object if no expected no is sent

TRUE if number of columns matches expected

FALSE if number of columns does not match expected
}
\description{
Validates the number of columns of an object

Returns number of columns if no parameter is used

If Verbose = F, returns dataframe of columns and bool if match
}
\seealso{
\code{\link{ncol}}
}
