% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindColNA.R
\name{FindColNA}
\alias{FindColNA}
\title{Identifies Number and Percentage of NAs By Column}
\usage{
FindColNA(x, verbose = TRUE)
}
\arguments{
\item{x}{A matrix-like R object}

\item{verbose}{Logical value to print results to console}
}
\value{
A dataframe of results
}
\description{
The NA column details including the column name, the count
of NA's for each column, and the percentage of each column those NA's
represent
}
\examples{
\dontrun{
#EXAMPLE
}
x = matrix(rnorm(100), ncol = 10)
FindColNA(x)
}
