% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RowDup.R
\name{RowDup}
\alias{RowDup}
\title{Identifies Duplicate Rows in Object}
\usage{
RowDup(x, verbose = TRUE)
}
\arguments{
\item{x}{A matrix-like R object}

\item{verbose}{Logical operator}
}
\value{
A dataframe of results
}
\description{
Identifiction of duplicated rows contained in the
data set. Displaying a table of the count of duplicate rows, the
percent of duplicated rows compared with the total rows, as well as
the count of unique rows.
}
\examples{
\dontrun{
x <- data.frame("X1" = c(1,2,3,1,2,3),
                "X2" = c(TRUE,FALSE,TRUE,TRUE,TRUE,FALSE),
                "X3"= c("a","b","c","a","b","c"))

RowDup(x)

value <- RowDup(x, verbose = FALSE)
}
}
\seealso{
\code{\link{unique}}

\code{\link{duplicated}}
}
