% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWNormality.R
\name{SWNormality}
\alias{SWNormality}
\title{Performing Shapiro-Wilk Normality Test}
\usage{
SWNormality(x, verbose = TRUE)
}
\arguments{
\item{x}{A matrix-like R object}

\item{verbose}{A logical indicator}
}
\value{
A dataframe of results
}
\description{
Performing Shapiro-Wilk Normality Test
}
\examples{
\dontrun{
x <- data.frame("Norm1" = c(11.2,8.2,10.0,27.3,14.5,26.4,4.2,15.2,14.5,26.7))

SWNormality(x)

value <- SWNormality(x, verbose = FALSE)
}
}
